backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
psycopg[binary]==3.2.9
alembic==1.16.4
pydantic-settings==2.10.1
email-validator==2.2.0
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
celery==5.5.3
redis==6.4.0
boto3==1.40.30
edge-tts==7.2.3
python-pptx==1.0.2
pymupdf==1.26.4
requests==2.33.1
